fix: resolve compilation errors for Flutter 3.44.4
- Fix DownloadStatus enum conflict between download_provider and best_downloader_service
- Add missing dispose() override to ConsoleLogOutput class
- Fix minLevel parameter to be nullable with default value
- Fix Duration(milliseconds) type error from math.pow() returning num
- Fix ErrorWidget.builder return type
- Fix Future.wait type annotation
- Add missing import for download_scheduler.dart
- Fix _openFile method to use DownloadItem instead of PersistedDownload
- Add _openFileByPath helper method
- Fix Icons.document_file to Icons.insert_drive_file
- Fix AndroidNotificationCategory.statusEvent to progress
- Change _totalSpeed and totalSpeedAll from int to double

+++ lib/services/best_downloader_service.dart
@@ -474,7 +474,7 @@ class BestDownloaderService {
         
         if (retryCount < MAX_RETRIES) {
           // Retry with exponential backoff
-          final delay = RETRY_BASE_DELAY * math.pow(2, retryCount);
+          final delay = (RETRY_BASE_DELAY * math.pow(2, retryCount)).toInt();
           _logger.info('BestDownloader', 'Retrying in ${delay}ms (attempt ${retryCount + 1})');
           
           await Future.delayed(Duration(milliseconds: delay));
